Refactor game logic by moving utility functions to the utils directory and adding type definitions for BoardProps and SquareProps
