+++ src/types/BoardProps.ts
