+++ src/utils/CalculateWinner.ts
